Added a note about www folder
This was my first custom lovelace card I added and I did not know about the need for www folder. I would have thought HACS would add it but the installation succeeded but did nothing. Even after a restart it was showing in HACS but did not exist on disk.

Ok still struggled next I found out I need to register the resources, you sort of mention in the manual setup  and the hacs popup it but again not clear. I cant see a hacsfiles folder but it did create the www/community/scheduler-card files. I doubt I have the information 100% correct here but at least it seems to be working now