Update sk.json
Items "excluded_items" and "additional_tasks" have a different format in my proposed localization compared to the main English one. This is due to the Slovak grammar which has 3 different forms of the words "item" and "task" when used in conjunction with numeral. 

An example with the word "task":
1 - "úloha"
2,3,4 - "úlohy"
5+ - "úloh"

Unfortunately I am not able to amend localize.ts file to accommodate this case as I am not familiar with TS.